Refactor Player's properties into PlayerData class
Refactored all Player's properties into a new class called PlayerData to improve the organization and readability of the code. Added PlayerSkills class to store player skills and levels. Adjusted references to Player's properties in out- and in-going packets as well as in the PlayerManager. Created methods for loading and saving PlayerData in json format. Corrected parameters names in SetSidebarInterfacePacket class's Add method  to reflect their actual function. In PlayerCommandPacket, added a 'level' command to allow setting skill levels for the player.