Improve error handling and logging for network stream
This commit introduces several enhancements to better handle network streams and increase error logging. Firstly, it has made changes to Check NetworkStream for its readability and writeability before any operation. It also throws an IOException if not enough bytes are read from the stream. Further, the PacketHandler class was refactored to better support state-based handling. The Player initialization now also includes a check if a player data file exists and creates it if not. Logging has been increased and improved across to better troubleshoot or understand what is happening during runtime. These changes increase the stability, readability, and debugging ability of the network layer for the game server.