Refactor game code for better user data handling
This change removes the PlayerData property from the Player class, and introduces it to the Client class. This decision was made to make the data more accessible in every relevant class where a client's data is needed, without going through the player class. In addition, all individual instances where the Username and Password were fetched directly have now been replaced with instances where they are fetched from the Data property of the Client.

Moreover, the target framework has been downgraded from net8.0 to net7.0 for better compatibility with older systems. Also, the PlayerUpdatePacket class has been refactored to receive Player input as a parameter in the constructor and use this as a readonly field, improving the readability of the code.

It is important to note that this commit also includes a logging adjustment in the ConnectionHandler and LogoutPacket where it changed the property access route for Username.