Add player options for trade and follow
This commit adds a new functionality to the player interaction options in the RuneRebirth2005 game. Two options are added: "Trade with" and "Follow". This is accomplished by creating a new "SetPlayerOptionsPacket" class in the Network/Packets/Outgoing directory, and adding calls to this new class in the PlayerManager. This update will enhance the user experience by enabling more player interactions.