"Implemented basic Melee combat system"
The commit introduces a basic Melee combat system into the game. The newly added CombatManager is now invoked in the main ServerEngine thus bringing it live. DamageInformation class is added to facilitate keeping track of damage info. The NPCs and Players in the game now have Weapons and MeleeCombat instances. The latter calculates attacks and decides what combat animation to play.

The "kill" and "unkill" player commands have been adjusted to reflect the changes. The 'Alive' property of NPCs is no longer toggled through this command. AnimationId property has been renamed to CurrentAnimation to improve clarity.

The "RegularWalkPacket" has a new assignment that nullifies the CombatFocus of a player to ensure that after making a move, the player isn't locked onto an enemy.

The incoming PlayerCommandPacket and outgoing PlayerUpdatePacket now handle the newly added combat features. Changes in the NPCUpdater class ensure that CurrentAnimation is properly reset and the combat processing is done accordingly. In the NPC class, a MeleeCombat instance is instantiated and other properties necessary for combat are also initialized.

This commit serves as the foundation for further combat-related feature development.