Add delayed attack task and improve combat system
Added a new feature named DelayedAttackTask that holds a list of tasks to be done after a specific number of ticks. This has been primarily added to improve the combat system where there was a need for a combatant, after hitting their target, to wait a tick before re-engaging.

In addition to the above, enabled an option to identify whether an entity is 'InCombat'. This will help in accurately portraying an entity's combat status and improve gameplay dynamics.

Made changes to how combat animations are decided for both player and NPC entities, with a focus on improving game accuracy and minimising potential errors.

Please note that these changes can influence overall gameplay mechanics and flow, and thus thorough testing may be required.